rename the directory
