+++ src/views/sign/ForgetPw.js
